Add sidekiq and use redis for action cable in development
sidekiq needs redis.

Don't use Active Job because this makes it sound like sidekiq and
Active Job don't get along well:
https://github.com/mperham/sidekiq/wiki/Active+Job

Also use redis for Action Cable since we have it. (`redis-cli monitor`
was handy for verifying that redis was indeed being used. Also the fact
that the Rails server crashed when redis was off and when and only when
a client tried to open a web socket.)