Implement minimal login functionality
Need to be able to identify users to prevent one user sending player
event messages as a different user's player.

I had some trouble with making sure that a user's authentication
credentials are available in Action Cable connections after logging in.
Normally users log in by storing a user id with the `session`
controller method, but that is not available in the connection class,
and `cookies.encrypted` and `cookies.signed` don't seem to access the same
data as `session`. Oh well, do it this way, at least for now.